Add ordering to nat changeset (#5036)
When querying the DB for nat entries, it is not sufficient to
filter for entries with a version_added and version_removed greater
than a specified value, then order. The sorted columns need to be
interleaved as well during the query, or we may accidentally skip
entries. This commit adds a view to the db that enables such a query.