Remove thing-flinger
Thing-flinger was used to deploy multinode omicron clusters running
on commodity hardware for early testing needs. This was primarily used
for early experimentation with trust-quorum secret sharing and other
preliminary multi-node experiments. It deployed via SSH and was never
intended to be used forever. It was a stop-gap that nobody has used in
a while now that we have real hardware with wicket and the a4x2 falcon
testbed.

Fixes #5034