[nexus] remove `services` queries & OMDB commands
Depends on #5033.

As described in #4947, we would like to remove the `services` table, as
its only remaining use is DNS propagation. PR #5033 changes DNS
propagation to no longer use the `services` table. Now that we're no
longer consuming this table, this commit removes the one method which
queries that table (`Datastore::services_list_kind`) and its two
remaining callers, the OMDB commands that query the table. I've also
removed the test for inserting and querying this table.

Note that I have *not* yet removed the code in RSS that actually
populates this table when the rack is set up. I thought it might be
desirable to still populate this data in case we have to roll back to a
previous version of Nexus that uses the `services` table. If this isn't
something we care about, I can remove that code as well, allowing us to
remove the `Datastore` methods for inserting to `services`.