[nexus-db-queries] remove uses of filter_target (#5052)
`filter_target` is required for partial indexes but is easy to misuse
for full indexes. To work around that, define an `as_partial_index`
method, switch to it, and ban general uses of `filter_target` via
clippy.

`as_partial_index` adds a `WHERE false` clause. See the comments
included in the PR for why that works.

The rest of the PR:

* Makes sure clippy's `disallowed_` lints are active. `-A clippy::style`
disables them.
* Adds some missing tests (thanks @jmpesp)
* Fixes a `RunnableQuery` that often did not return anything, replacing
it with `RunnableQueryNoReturn`.