`IncompleteNetworkInterface`: allow user-specified explicit `slot` (#5065)
This makes a several minor changes to plumb slots through:

* `IncompleteNetworkInterface` now stores an optional slot, just like it
stores an optional IP/MAC address
* In `network_interfaces::InsertQuery`, if the incoming slot is set, we
use it directly instead of running the `NextItem`-based subquery
* Adds a partial index to ensure uniqueness of a slot within a single
`parent_id` (I believe this is correct, but would love confirmation from
someone more familiar!)
* `IncompleteNetworkInterface::new_service()` now takes a _non-optional_
IP, MAC address, and slot. This matches how it was called in all
non-test code.
* Tweaked the Nexus internal API used for RSS handoff to include the
slot in the description of NICs.

This is a partial fix for #5056, and should produce correct behavior on
new systems that run through RSS, even without a fix for #5055 (because
we bypass `NextItem` altogether with this change). In particular, I
think this should unblock testing of #5045 on madrid / testbeds. It does
not address the already-recorded-NICs-with-incorrect-slots on systems
like dogfood; I'll take care of that in a subsequent PR.