Blueprint execution: Add dataset records for Crucible zones (#5143)
On each invocation, for every Crucible zone present in the blueprint
(all of which have already successfully been sent to sled-agent by this
point), we attempt to insert a `dataset` record for that zone, but only
if a record does not already exist. The datasets themselves are created
by sled-agent when we send the zone configs.

Fixes #5118.