Enable runtime toggling of RPWs
Background tasks in Nexus are initialized before Nexus itself. This
appears to introduce a possibility of a race in the RSS workflow:
* A rack is initialized via RSS, but its settings are not yet in
  the database
* The RPWs start running before Nexus is fully initialized
* The RPW performs changes based on missing or partial information
* The RPW's changes possibly interferes with Nexus's ability to
  fully initialize, or temporarily disrupts a user's connectivity
  to Nexus until the conflict is resolved

A strategy introduced in this commit is flag set in the db to enable
RPWs at runtime. This way we can make the RPW wait until nexus has
enabled it.