Expose uninitialized sleds and "add sled" via internal API and omdb (#5204)
This will let us drive the add sled process from a support context.
(Prior to this change adding a sled required interaction with the
external API.)

I want to test this before merging, but it's small enough that if
there's anything that needs tweaking it's unlikely to affect any review
comments. One change is that we had two `Baseboard` structs that were
identical except for a minor different in field names (`{serial,part}`
vs `{serial_number,part_number}`). I removed the internal one and
replaced it with the external one to avoid unnecessary churn in the
public API, but if I'm missing something intentional that led to having
both, please say so.

Closes #5133.