Groundwork for oximeter instance vCPU metrics (#4900)
- Adds the silo and project IDs to the instance-ensure request from
Nexus to the sled-agent. These are used as fields on the
instance-related statistics. This metadata is currently unused, but will
be forwarded to Propolis in the instance-ensure request once the server
is updated to accept it.
- Defines a `VirtualMachine` oximeter target and `VcpuUsage` metric. The
latter has a `state` field which corresponds to the named kstats
published by the hypervisor that accumulate the time spent in a number
of vCPU microstates. The combination of these should allow us to
aggregate or break down vCPU usage by silo, project, instance, vCPU ID,
and CPU state. Adds some simple mocks and tests for these.
- Adds more fine-grained feature flags to the `oximeter-instruments`
crate.