`omdb nexus blueprints target set`: take an `enabled` setting
I made this a required argument because I'm nervous to default it to
`Enabled` when we've talked so much about manually setting a disabled
blueprint. If folks feel strongly this should be an optional
`--enabled={true,false,inherit}` defaulting to `true` instead, I
wouldn't put up a fight.

Addresses the first part of #5210.