Unquarantine clickhouse binary on mac (#5223)
After #5127 (I assume, because this problem came up immediately after
that) macOS Gatekeeper is putting the `clickhouse` binary in quarantine
after download, which means we can't execute it unless we take it out of
quarantine. This is apparently a new thing, as [this
doc](https://github.com/ClickHouse/clickhouse-docs/blob/08d7a329d/knowledgebase/fix-developer-verification-error-in-macos.md)
about how to do that was only [added Jan
9](https://github.com/ClickHouse/clickhouse-docs/pull/1835).

Here I am doing it in the simplest way possible: if mac, then remove
quarantine attr. If anyone has a better way, I'm all ears.