Teach `omdb db network list-eips` and `DataStore::vpc_resolve_to_sleds` about blueprints (#5202)
While working on #4886, I ran into a couple places that expect all
services to have rows in the `service` table, but Reconfigurator does
not add new rows to `service` when it spins up new services (see #4947
for rationale). This PR teaches these two users of `service` to _also_
check the current target blueprint when checking for services.

I also removed `LookupPath::service_id` and `lookup_resource! { ...
Service ... }`; the only user of these was `omdb`, and I think their
existence is misleading in the post-`service`-table world we're moving
toward.