Add LLDP (#5196)
This adds the `lldp` daemon and CLI to the switch zone and updates
`sled-agent` to do the basic configuration and enablement of the
service.

On its own, this change will not cause the switch zone to advertise or
receive `lldp` packets, as no interfaces are configured. That can be
done using the CLI in the switch zone for testing on `dogfood` and/or
the colo. Adding runtime configuration by `nexus` will be follow-on
work.