Split out a `sled-hardware-types` crate (#5245)
On main as of #5158, we unexpectedly get some binaries depending on
`libnvme` that shouldn't. In release builds:

```
installinator
omicron-dev
omicron-package
services-ledger-check-migrate
sled-agent
sled-agent-sim
wicketd
```

and in debug builds, all of the above plus `omdb`. We don't really care
about this for binaries that don't run on the rack, so stripping the
list down to binaries we do care about:

```
installinator
omdb (debug only)
sled-agent
wicketd
```

It's correct and expected that installinator and sled-agent depend on
libnvme, but omdb shouldn't (and doesn't in release), and wicketd _must
not_, as libnvme isn't available in the switch zone.

This PR fixes that incorrect dependency by splitting the parts of the
`sled-hardware` crate that wicketd and omdb depended on (directly or
transitively) into a new `sled-hardware-types` crate. On this PR, we are
left with only the following needing libnvme, in both debug and release:

```
installinator
omicron-dev
omicron-package
services-ledger-check-migrate
sled-agent
sled-agent-sim
```

I assume with a bit more work we could trim out everything but
`installinator` and `sled-agent`, and that might be worth doing, but
we'd like to land this ASAP so as to not break any updates performed off
of `main`. Separately, we could also imagine a CI check that we don't
have unexpected library dependencies present in the final binaries;
@papertigers is going to work on that.