Add internal API and omdb commands to toggle current target blueprint's enabled bit (#5231)
I chose `PATCH` for the new endpoint; I think that's correct? Although
it's the first use of `PATCH` in omicron AFAICT so maybe it should just
be `PUT`?

There are a couple related-but-not-required changes as a part of this:

* `omdb nexus blueprints list` now has an `ENA: yes/no` column that is
only populated for the current target
* We activate the blueprint execution bg task when setting an enabled
target or changing the current target to be enabled. (The latter might
be superfluous if the target was already enabled, but that's fine.)

I tried this locally against `omicron-dev run-all` and it all seemed to
check out. If folks would be more comfortable with a test on a real
system before merging I can do that; otherwise I'll test it next time I
take a lap for some other reason.

Closes #5210.