RSS <-> Nexus handoff: Set an initial (disabled) target blueprint (#5244)
I would like to test that after an RSS handoff, the initial blueprint we
created in RSS and set in the Nexus handoff matches the blueprint we
would generate from the first inventory collections. I can do that by
hand on `madrid`, but I'm not sure there's a good way to do that in an
automated test. Is this something I could lean on a4x2 to check?

This does not modify the `services` table or the `services` field of the
RSS handoff. Removing those will come in a followup PR that will need an
accompanying note for deployed systems (i.e., instructions to set a
blueprint so the current users of `services` will continue to function
by checking the current target blueprint).

Closes #5222.