[nexus] Add zpool to inventory (#5249)
This PR adds zpools to the inventory, and updates the existing
`omicron.public.zpool` table to only include aspects of the zpool which
are fully controlled by Nexus.

This PR is intended to solve a problem introduced by
https://github.com/oxidecomputer/omicron/pull/5172 . In that PR,
physical disk and zpool provisioning will be decided by Nexus, and sent
to the Sled Agent as a request. As a result, the `total_size` value of
zpools is not known when a pool is provisioned. Here, we add zpools to
the inventory, and JOIN with that table to access the "total_size" of
the zpool if it is known.

Additionally, this PR adds a test to validate that the database
(appropriately) throws "Insufficient capacity" errors if the Zpool
exists, but has not appeared in the inventory.