Rpws for all networking (#4822)
Overview
---
This PR ensures the rest of our `dpd` configuration is covered by a RPW
to help recover state in the event of `dendrite` crashing, the switch
zone restarting / being replaced, or the sled restarting. This is
accomplished via a background task in Nexus that periodically ensures
`dpd` is up to date with the tables in Nexus. The tradeoffs of this
design is that we don't track versioning and reconcile the entire state
every time, but since the actual number of ports will never be that high
(relative to something like NAT entries) the tradeoff of less efficiency
for much greater simplicity seems to make sense today, and it requires
much less rework in Nexus and Dendrite should we choose to replace this
strategy down the road.

Tasks
---
- [x] Ensure that Service Zones configured during rss, cold boot, and
nexus have their NAT entries added to the NAT RPW table (extracted into
#4857)
- [x] Create background task that periodically reconciles switch port
configuration for dendrite instances
- [x] Move switch zone uplink SMF property updates to RPW
- [x] Move routing updates (via mg) to RPW
    - [x] Static Routing
    - [x] BGP
- [x] Move bootstore updates to RPW
- [x] Move loopback address management to RPW
- [x] Move Nexus-side switch zone service on-demand lookups as outlined
in #5092

Verifications Performed
---
- [x] Basic instance deployment
- [x] Loopback Address Creation
- [x] BGP configuration (a4x2)
- [ ] BGP configuration modification (a4x2)
- [x] Static routing
- [x] Static routing configuration modification

Related
---
Closes #4715
Closes #4650

Depends on https://github.com/oxidecomputer/dendrite/pull/838

---------

Co-authored-by: Levon Tarver <levon@oxidecomputer.com>