Current IP pool utilization endpoint (#5258)
Closes #5189


The strangest thing here is that IPv6 `allocated` and `capacity` are
both u128 that we put in the JSON spec as `"type": "string", "format":
"uint128"`, where `uint128` is a custom format our client generators
will have to look for and handle. `capacity` is huge because IPv6 ranges
are huge. IPv4 allocated and capacity are `u32`.

~~The strangest thing here is the fact that we have to handle the
unlikely possibility that an IPv6 range is too large for us to serialize
its length in a JSON response. IPv6 range size is a `u128`, and we seem
to be capping our numbers in API responses at `u32`. My proposal for
handling this is simply to make the field nullable in the response, and
we make clear that a null value means it was too large to serialize. I
think this is fine because this should only happen if the operator does
something really bad, and it is well understood that the operator can
mess up networking with a misconfiguration. Another approach would be to
always return the size as a string, but I think given the unlikelihood
of the number being too large (and the fact that it's pathological if it
is) means returning a numeric string puts undue burden on clients to
handle it.~~