Restore instance tests to #5172
Restore instance tests added in #4691 so that they work with the latest
updates in #5172. The tests now use the new `StorageManagerTestHarness`
from #5172 instead of the since removed `FakeStorageManager`.
Additionally, the dns server code from `sled-agent/src/fakes/nexus` is
used rather than the reimplementation in #4691. This shrinks the amount
of code added with these tests.