More review feedback
This Implements two new limits on query complexity:

- Maximum amount by which we allow an alignment query to upsample the
  data, and
- Maximum number of total measurement values fetched from ClickHouse

The first is easy to check during the `align` table operation, and I've
added tests for it.

The second is checked by continually updating the total number of rows
we receive from the database on each measurement query. This also
changes the way we _ensure_ that limit. Previously, we ran the
measurements query twice: once to check that it didn't return too much
data, and the second to fetch the actual results.

That's an obvious TOCTOU, and inefficient, but it was simple. Instead,
this commit changes way we do that check, by applying a limit clause to
the next measurement query. The limit we use is one more than the
remainder in our allotment. That way, we can tell if the query is too
big, without running the query multiple times or returning far more data
than we want to allow. We don't know how _much_ bigger the actual result
set is than what we want, but we do know that it is bigger.