Correctly implement group_by via timestamp matching
During early development, I implemented a short-cut for group_by
operations, which grouped values by _index_ not timestamp. That worked
because the inputs were aligned and usually from timeseries with exactly
overlapping points. It's not really correct, however. We _want_ to group
data points in a timeseries both by their group identity and _also_ by
lining up those that occur at the same time.

This commit implements correct grouping by finding matching output
timestamps in aligned input timeseries, and averaging or summing only
those at the same time. Care is taken to not include missing points when
computing the averages (or sums).

This also found a bug in the alignment code. That used the
`Iterator::sum` implementation when averaging values within a time
window. That method doesn't return `None` if there is no data, it always
returns the zero value for the input. That caused the grouping code to
convert missing values to zeros, which when you're summing is fine, but
not when averaging.