[reconfigurator] replace Blueprint::zones_in_service with a disposition enum next to each zone (#5238)
This greatly simplifies the handling of the expunged disposition in
#5211. It also is more understandable, and less prone to making
mistakes.

This PR only changes the in-memory representation of what is currently
`zones_in_service`. I've skipped doing the DB migration in this PR
because it'll be easier to just do it wholesale in #5211.

There are some small differences in diff output that I think make sense.
But I took a bigger swing at it in #5270, which depends on this.