WIP: Allow control of tests run in VMs on the host
This is a reworking of `falcon_runner.rs` that splits the work between
a `FalconTest` harness that allows direct manipulation of the runner
so that the full falcon API is available, while also wrapping the bash
nonsense to run a test and get the result back.

The harness is minimal at this point, and purposefully so, to allow for complete
flexibility. This allows the user to launch multiple VMs, run arbitrary tests,
mount arbitrary data directories into VMs, etc... We can provide more batteries
included, perhaps with escape hatches, and we can also simplify further
with attribute macros for the batteries included parts.