[nexus] move UuidRng code out to a common crate, make CollectionBuilder use it (#5341)
In #5270, we need determinism not just from blueprints but also
collections. So move the UuidRng into a common place.

As part of that, I also decided to make it its own crate and write some
documentation about it, making it more generic along the way. I think
this should be a pretty clean representation of what this is trying to do.