Add basic Oximeter query language (#5273)
- Add basic grammar for an Oximeter query language. Includes support for
numeric, string, boolean, UUID, timestamp, IP address, and duration
literals. Queries are constructed in a pipeline of "table operations",
each of which operates on a set of timeseries and produces another set.
- Implement temporal alignment, currently supporting one method that
generates output samples from the mean of the inputs over the alignment
period.
- Add basic subquery support, for fetching multiple timeseries and
joining them
- Implement filtering on fields and timestamps, both in the DB as much
as possible, and the query pipeline; and implement filtering on data
values in code.
- Implement group-by support, where we can currently reduce values
within a group by summing or computing the mean.
- Add public Nexus API endpoints for listing timeseries schema, and
running an OxQL query. Both are currently restricted to fleet readers,
until a more thorough authz process is fleshed out.
- This also reorganizes the internals of the `oximeter_db::client`
module, which were starting to get too unwieldy and littered with
conditional compilation directives.